Just replace the old pattern generator
