+++ platforms/allwinner-d1/core/src/drivers/mod.rs
@@ -1,4 +1,3 @@
 pub mod sharp_display;
 pub mod spim;
 pub mod uart;
-pub mod emb_display;
